+++ flippy/flippy.py
@@ -1,7 +1,19 @@
+from dataclasses import dataclass
 from flippy.core import BaseBackend, FeatureName, Gate
 from flippy.exceptions import FeatureNotFound
 
 ACTOR_IF_NO_TARGET = "anonymous"
+
+
+@dataclass
+class FeatureState:
+    key: str
+    boolean: bool | None
+    actors: list[str]
+    groups: list[str]
+    percent_actors: int | None
+    percent_time: int | None
+
 
 class Flippy:
     def __init__(self, backend: BaseBackend):
@@ -26,13 +38,20 @@ class Flippy:
         else:
             actor = self._to_flipper_id(target)
 
-        return any([
+        if any([
             f.actors_gate.is_open(actor, feature),
             f.groups_gate.is_open(actor, feature),
             f.percentage_of_actors_gate.is_open(actor, feature),
             f.percentage_of_time_gate.is_open(actor, feature),
             f.expression_gate.is_open(actor, feature),
-        ])
+        ]):
+            return True
+        
+        # TODO: special check whether actor is a member of an enabled group
+        # (using the Django authentication system)
+        ...
+        
+        return False
     
     def create(self, feature: FeatureName) -> bool:
         return self._backend.add(feature)
@@ -46,6 +65,17 @@ class Flippy:
             return True
         except FeatureNotFound:
             return False
+    
+    def get_feature_state(self, feature: FeatureName) -> FeatureState:
+        f = self._backend.get(feature)
+        return FeatureState(
+            key=f.key,
+            boolean=f.boolean_gate.value,
+            actors=f.actors_gate.value,
+            groups=f.groups_gate.value,
+            percent_actors=f.percentage_of_actors_gate.value,
+            percent_time=f.percentage_of_time_gate.value,
+        )
     
     def enable(self, feature: FeatureName) -> None:
         try: