+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # In flight
+
+## Major updates
+- Added a Flippy middleware (for easy access from view functions)
+- Added a Flippy context processor (for easy access from templates)
+- Demonstrated how to use Flippy with a demo app
+
+## Minor updates
 - Make clear the relationship to Flipper Cloud (i.e., there is none).
 - Remove Markdown from PyPI description metadata.